moved the function buildDialog to the father's class and removed it from all the dialogs
